Implementacao do keyholder para recuperar o id do cliente inserido na base.

+++ src/main/java/br/com/estudos/crud/repository/ClienteRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ClienteRepository {
 
-    int insertClient(ClienteRequest request);
+    Long insertClient(ClienteRequest request);
 
     ClienteDto findByCpf(String cpf);
