+++ src/main/java/br/com/estudos/crud/service/CadastroClienteService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CadastroClienteService {
 
-    int cadastrar(ClienteRequest request);
+    Long cadastrar(ClienteRequest request);
 
     ClienteDto findByCpf(String cpf);
